Consolidate InlineTemplateEdit into Template component
Template.tsx now handles both /template/:templateId and
/template/:templateId/from/:repeatableId routes by detecting the
optional repeatableId param. When editing from a repeatable context,
shows the repeatable's checkbox values in preview and migrates it
to the new template version on save.
