+++ TODO
@@ -1,8 +1,14 @@
 Quick list
 
+inline edit:
+- fix back button behaviour
+- generalise migrations
+
+WYSIWYG editor.
+
 Revisit typing. We should have internal types not bound to PouchDB structures that mean the slug against the repeatable are cleanly typed with no ambiguiity.
 
-Button to edit template on repeatable view. Will edit existing repeatables. This has downsides because repeatable data is just an array of booleans, not a map. Think about that but probably don't care for now.
+Move to pnpm because I use it everywhere else.
 
 Spend a pomo looking into shorter and less ugly variations on raw uuids to clean up URLs etc.
